Feature - Add RabbitMQ and MassTransit

+++ Profiles.Business/Implementations/Repositories/DoctorSummaryRepository.cs
@@ -25,7 +25,7 @@ namespace Profiles.Business.Implementations.Repositories
             parameters.Add("Id", dto.Id, DbType.Guid);
             parameters.Add("SpecializationName", dto.SpecializationName, DbType.String);
             parameters.Add("OfficeAddress", dto.OfficeAddress, DbType.String);
-            parameters.Add("Status", (int)dto.Status, DbType.Int32);
+            parameters.Add("Status", dto.Status, DbType.Int32);
 
             using (var connection = _db.CreateConnection())
             {
@@ -47,7 +47,7 @@ namespace Profiles.Business.Implementations.Repositories
             parameters.Add("Id", id, DbType.Guid);
             parameters.Add("SpecializationName", dto.SpecializationName, DbType.String);
             parameters.Add("OfficeAddress", dto.OfficeAddress, DbType.String);
-            parameters.Add("Status", (int)dto.Status, DbType.Int32);
+            parameters.Add("Status", dto.Status, DbType.Int32);
 
             using (var connection = _db.CreateConnection())
             {
